+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package rosros
 ------------------
 * add rosros.wait_for_publisher(), wait_for_subscriber() and wait_for_service()
 * fix rosros.get_services() raising error on node-parameter in ROS2
+* fix rosros.get_param_names() returning params dictionary instead of list
 * fix rosros.set_param() raising error on no descriptor in ROS2
 
 0.2.1 (2023-12-06)