+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog for package rosros
 * fix rosros.set_param() raising error on no descriptor in ROS2
 * fix util.get_arity() and wrap_arity() raising for built-in functions
 * fix rosros.spin_once() raising error on negative timeout in ROS1
+* fix api.get_message_type() not returning value for service in ROS1
 * fix throttling log messages again
 
 0.2.1 (2023-12-06)